* 添加进度条 * 添加节拍模型 * 修复之前的一些UI与逻辑

+++ BeatDemo/BeatDemo/Class/DataModel/PassNote.swift
@@ -27,9 +27,9 @@ class PassNote: NSObject {
     }
     
     /// 进入拍子
-    var enterBeat: UInt8!
+    var enterBeat: Int!
     /// 退出拍子
-    var exitBeat: UInt8!
+    var exitBeat: Int!
     
     init(midiNoteNumber: UInt8,
          enterTime: double_t,
